// packages/ajv-human-errors/src/util.d.ts
export declare const capitalize: (s: string | null) => string | null;
export declare const pluralize: (s: string, num: number) => string;
export declare const jsonPath: (s: string) => string;
export declare const humanizePath: (s: string) => string;
export declare const indefiniteArticle: (s: string) => "a" | "an";
export declare const humanizeTypeOf: (value: any) => string;
export declare const humanizeList: (arr: string[], conjunction?: string) => string;